Refactor to use assets background and foreground painters
