+++ features/il_display_assets_map/lib/src/widgets/assets_painter.dart
@@ -2,14 +2,42 @@ import 'package:flutter/material.dart';
 import 'package:il_core/il_entities.dart';
 import 'package:il_core/il_ui_helpers.dart';
 
-class AssetsPainter extends CustomPainter {
+class AssetsBackgroundPainter extends CustomPainter {
+  final FloorMap floorMap;
+  final IFloorMapImageRenderer imageRenderer;
+
+  late FloorMapRenderer floorMapRenderer;
+
+  AssetsBackgroundPainter({
+    required this.floorMap,
+    required this.imageRenderer,
+  }) {
+    floorMapRenderer = FloorMapRenderer();
+  }
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    imageRenderer.draw(canvas);
+    floorMapRenderer.drawZones(canvas, floorMap);
+  }
+
+  @override
+  bool shouldRepaint(CustomPainter oldDelegate) {
+    if (oldDelegate is AssetsBackgroundPainter) {
+      return floorMap != oldDelegate.floorMap;
+    }
+    return false;
+  }
+}
+
+class AssetsForegroundPainter extends CustomPainter {
   final Matrix4 transform;
   final FloorMap floorMap;
   final List<Asset> assets;
 
   late FloorMapRenderer floorMapRenderer;
 
-  AssetsPainter({
+  AssetsForegroundPainter({
     required this.transform,
     required this.floorMap,
     required this.assets,
